BasicPainter class divided into three separate painters. (basic_edges_painter.py - not active by default, basic_selection_painter.py - active by default, basic_fill_painter.py - active by default). Added Painters submenu in tools menu, where each painter have checkbox menu for processing the data and for painting
