--- src/modules/painters/a_painterbasic/__init__.py
@@ -1,11 +1,0 @@
-import sys
-import os
-
-sys.path.append(os.path.dirname(__file__))
-
-from a_painterbasic.basicpainter import BasicPainter
-
-
-
-def createPainter():
-    return BasicPainter()
