+++ src/modules/painters/dir_basic_painter/glhelp.py
