+++ src/modules/painters/dir_basic_painter/glvertdatasforhaders.py
@@ -9,10 +9,10 @@ from PySide6.QtOpenGL import QOpenGLBuffer, QOpenGLVertexArrayObject
 from PySide6.QtGui import QOpenGLContext
 from shiboken6 import VoidPtr
 
-from a_painterbasic.glhelp import GLDataType
+from dir_basic_painter.glhelp import GLDataType
 
-from a_painterbasic.glvertdata import VertDataCollector
-from a_painterbasic.glvertdata import VertDataSingleChannel
+from dir_basic_painter.glvertdata import VertDataCollector
+from dir_basic_painter.glvertdata import VertDataSingleChannel
 
 
 # class that use Coordinates and normals
